+++ Backend/eewms/src/main/java/com/eewms/dto/warehouseReceipt/WarehouseReceiptItemDTO.java
@@ -15,7 +15,7 @@ public class WarehouseReceiptItemDTO {
     private String productName;
     private BigDecimal price;        // Giá nhập hàng thực tế
     private Integer actualQuantity;
-    private Long warehouseId;
+    private Integer warehouseId;
     private String warehouseName;
     private Integer contractQuantity;
 }
